+++ lib/simply_stored/couch.rb
@@ -274,5 +274,18 @@ module SimplyStored
       end
       
     end
+    
+    def self.delete_all_design_documents(database)
+      db = CouchRest.database(database)
+      db.info # ensure DB exists
+      design_docs = CouchRest.get("#{database}/_all_docs?startkey=%22_design%22&endkey=%22_design0%22")['rows'].map do |row|
+        [row['id'], row['value']['rev']]
+      end
+      design_docs.each do |doc_id, rev|
+        db.delete_doc({'_id' => doc_id, '_rev' => rev})
+      end
+      design_docs.size
+    end
+    
   end
 end